pyproject.toml
[tool.poetry]
name = "agents-capstone"
version = "0.1.0"
description = "Google ADK based agents application for property maintenance triage"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
google-genai = "^1.0.0"
google-adk = {extras = ["a2a"], version = "^1.19.0"}
python-dotenv = "^1.2.1"
pandas = "^2.3.3"
aiosqlite = "^0.21.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-asyncio = "^0.21.0"
black = "^23.0.0"
flake8 = "^6.0.0"

[[tool.poetry.packages]]
include = "src"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"